On aura besoin de graaaands pochoirs pour les obstacles grossis

+++ pc/src/pathfinding/Pathfinding.java
@@ -41,7 +41,7 @@ public class Pathfinding implements Service
 		config = requestedConfig;
 		log = requestedLog;
 		centimetresParCases = requestedCentimetresParCases;
-		map = new Grid2DSpace(new IntPair(300/centimetresParCases, 200/centimetresParCases), table);
+		map = new Grid2DSpace(new IntPair((int)((float)300.0f/centimetresParCases), (int)((float)200.0f/centimetresParCases)), table);
 		solver = new AStar(map, new IntPair(0,0), new IntPair(0,0));
 		output = new ArrayList<Vec2>();
 		result = new ArrayList<IntPair>();
